Added Dev Page + Large Update
Add CSS to all the things.
Made all of the parts/games the same across the board. (Only Minecraft and Elden Ring)
Made it so when you save a build in the hardware page, it updates the MyBuild on the right side.
Made both the admin/dev dashboard only show one product at a time (it just felt a bit weird to have them all at once from the beginning when there was no way to get back to that)
Added Several Dropdown boxes to Admin Page
Added Dev Page
Updated FAQ Page
Added the 3 logins from the admin page to the login (passwords are 1234)