Implement search for one game in database via name.
For some reason trying to search via appID for games doesn't work.
I suspect this is due to the input text-box on the admin page sending
in strings to search by, meanwhile the appID datatype is a numeric.
Too lazy at the moment to figure this out and implement with such
a close deadline so for now game names will be primary key for the
games collection (both steam and local lists).